Move apollo to apollotracing and add py36 fallback for perf_counter_ns
